+++ tests/testthat/test-job.R
@@ -11,7 +11,7 @@ test_that("job_data() calls are correct", {
   exp_data_frame <- utils::read.csv(cdc_file, stringsAsFactors = FALSE, colClasses = 'character')  # "NA" -> NA
   mock_reauthenticate <- mock(exp_data_frame)
   testthat::with_mock("zoltr::get_resource" = mock_reauthenticate, {
-    act_data_frame <- job_data(zoltar_connection, "http://example.com/api/job/2/")
+    act_data_frame <- job_data(zoltar_connection, "http://example.com/api/job/2/", TRUE)
     expect_equal(length(mock_calls(mock_reauthenticate)), 1)
     expect_equal(mock_args(mock_reauthenticate)[[1]][[2]], "http://example.com/api/job/2/data/")
     expect_is(act_data_frame, "data.frame")
@@ -20,7 +20,7 @@ test_that("job_data() calls are correct", {
 })
 
 
-test_that("job_data() column types are correct for CSV data", {  # JSON data tested many other places
+test_that("job_data() column types are correct for forecast CSV data", {  # JSON data tested many other places
   zoltar_connection <- new_connection("http://example.com")
   mock_reauthenticate <- mock()
   mock_from_json <- mock()
@@ -32,10 +32,46 @@ test_that("job_data() column types are correct for CSV data", {  # JSON data tes
           to_return(
             status = 200,
             headers = list('Content-Type' = 'text/csv'))
-        job_data(zoltar_connection, "http://example.com/api/job/2/")
+        job_data(zoltar_connection, "http://example.com/api/job/2/", TRUE)
         expect_equal(length(mock_calls(mock_from_json)), 0)
         expect_equal(length(mock_calls(mock_read_csv)), 1)
         expect_equal(mock_args(mock_read_csv)[[1]]$col_types, "cDcccc?????????")
+        # note: we do not test the column types of the actual data returned by job_data b/c we trust that
+        # readr::read_csv correctly applies col_types
+      })
+    })
+  })
+})
+
+
+test_that("job_data() column types are correct for score CSV data", {  # JSON data tested many other places
+  zoltar_connection <- new_connection("http://example.com")
+  mock_reauthenticate <- mock()
+  mock_from_json <- mock()
+  mock_read_csv <- mock()
+  testthat::with_mock("zoltr::re_authenticate_if_necessary" = mock_reauthenticate, {
+    testthat::with_mock("jsonlite::fromJSON" = mock_from_json, {
+      testthat::with_mock("readr::read_csv" = mock_read_csv, {
+        webmockr::stub_request("get", uri = "http://example.com/api/job/2/data/") %>%
+          to_return(
+            status = 200,
+            headers = list('Content-Type' = 'text/csv'))
+        job_data(zoltar_connection, "http://example.com/api/job/2/", FALSE)
+        expect_equal(length(mock_calls(mock_from_json)), 0)
+        expect_equal(length(mock_calls(mock_read_csv)), 1)
+
+        score_cols <- readr::cols(
+          .default = readr::col_double(),  # scores
+          model = readr::col_character(),
+          timezero = readr::col_date(format = ""),
+          season = readr::col_character(),
+          unit = readr::col_character(),
+          target = readr::col_character(),
+          log_single_bin = readr::col_logical(),
+          log_multi_bin = readr::col_logical(),
+          pit = readr::col_logical()
+        )
+        expect_equal(mock_args(mock_read_csv)[[1]]$col_types, score_cols)
         # note: we do not test the column types of the actual data returned by job_data b/c we trust that
         # readr::read_csv correctly applies col_types
       })